Fuzz test mmr (#26)

+++ src/trie/Bytes.sol
@@ -2,7 +2,7 @@ pragma solidity ^0.8.17;
 
 // SPDX-License-Identifier: Apache2
 
-import { Memory } from "./Memory.sol";
+import {Memory} from "./Memory.sol";
 
 struct ByteSlice {
     bytes data;
@@ -21,11 +21,11 @@ library Bytes {
         if (self.length != other.length) {
             return false;
         }
-        uint addr;
-        uint addr2;
+        uint256 addr;
+        uint256 addr2;
         assembly {
-            addr := add(self, /*BYTES_HEADER_SIZE*/32)
-            addr2 := add(other, /*BYTES_HEADER_SIZE*/32)
+            addr := add(self, /*BYTES_HEADER_SIZE*/ 32)
+            addr2 := add(other, /*BYTES_HEADER_SIZE*/ 32)
         }
         equal = Memory.equals(addr, addr2, self.length);
     }
@@ -46,10 +46,7 @@ library Bytes {
     // Requires that:
     //  - 'startIndex + len <= self.length'
     // The length of the substring is: 'len'
-    function read(
-        ByteSlice memory self,
-        uint256 len
-    ) internal pure returns (bytes memory) {
+    function read(ByteSlice memory self, uint256 len) internal pure returns (bytes memory) {
         require(self.offset + len <= self.data.length);
         if (len == 0) {
             return "";
@@ -64,11 +61,7 @@ library Bytes {
     // Returns the new copy.
     // Requires that 'startIndex <= self.length'
     // The length of the substring is: 'self.length - startIndex'
-    function substr(bytes memory self, uint256 startIndex)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function substr(bytes memory self, uint256 startIndex) internal pure returns (bytes memory) {
         require(startIndex <= self.length);
         uint256 len = self.length - startIndex;
         uint256 addr = Memory.dataPtr(self);
@@ -80,11 +73,7 @@ library Bytes {
     // Requires that:
     //  - 'startIndex + len <= self.length'
     // The length of the substring is: 'len'
-    function substr(
-        bytes memory self,
-        uint256 startIndex,
-        uint256 len
-    ) internal pure returns (bytes memory) {
+    function substr(bytes memory self, uint256 startIndex, uint256 len) internal pure returns (bytes memory) {
         require(startIndex + len <= self.length);
         if (len == 0) {
             return "";
@@ -97,11 +86,7 @@ library Bytes {
     // Returns the concatenated arrays:
     //  [self[0], self[1], ... , self[self.length - 1], other[0], other[1], ... , other[other.length - 1]]
     // The length of the new array is 'self.length + other.length'
-    function concat(bytes memory self, bytes memory other)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function concat(bytes memory self, bytes memory other) internal pure returns (bytes memory) {
         bytes memory ret = new bytes(self.length + other.length);
         uint256 src;
         uint256 srcLen;
@@ -110,49 +95,33 @@ library Bytes {
         uint256 src2Len;
         (src2, src2Len) = Memory.fromBytes(other);
         uint256 dest;
-        (dest, ) = Memory.fromBytes(ret);
+        (dest,) = Memory.fromBytes(ret);
         uint256 dest2 = dest + srcLen;
         Memory.copy(src, dest, srcLen);
         Memory.copy(src2, dest2, src2Len);
         return ret;
     }
 
-    function toBytes32(bytes memory self)
-        internal
-        pure
-        returns (bytes32 out)
-    {
+    function toBytes32(bytes memory self) internal pure returns (bytes32 out) {
         require(self.length >= 32, "Bytes:: toBytes32: data is to short.");
         assembly {
             out := mload(add(self, 32))
         }
     }
 
-    function toBytes16(bytes memory self, uint256 offset)
-        internal
-        pure
-        returns (bytes16 out)
-    {
-        for (uint i = 0; i < 16; i++) {
+    function toBytes16(bytes memory self, uint256 offset) internal pure returns (bytes16 out) {
+        for (uint256 i = 0; i < 16; i++) {
             out |= bytes16(bytes1(self[offset + i]) & 0xFF) >> (i * 8);
         }
     }
 
-    function toBytes8(bytes memory self, uint256 offset)
-        internal
-        pure
-        returns (bytes8 out)
-    {
-        for (uint i = 0; i < 8; i++) {
+    function toBytes8(bytes memory self, uint256 offset) internal pure returns (bytes8 out) {
+        for (uint256 i = 0; i < 8; i++) {
             out |= bytes8(bytes1(self[offset + i]) & 0xFF) >> (i * 8);
         }
     }
 
-    function toBytes4(bytes memory self, uint256 offset)
-        internal
-        pure
-        returns (bytes4)
-    {
+    function toBytes4(bytes memory self, uint256 offset) internal pure returns (bytes4) {
         bytes4 out;
 
         for (uint256 i = 0; i < 4; i++) {
@@ -161,11 +130,7 @@ library Bytes {
         return out;
     }
 
-    function toBytes2(bytes memory self, uint256 offset)
-        internal
-        pure
-        returns (bytes2)
-    {
+    function toBytes2(bytes memory self, uint256 offset) internal pure returns (bytes2) {
         bytes2 out;
 
         for (uint256 i = 0; i < 2; i++) {
@@ -175,10 +140,10 @@ library Bytes {
     }
 
     function removeLeadingZero(bytes memory data) internal pure returns (bytes memory) {
-        uint length = data.length;
+        uint256 length = data.length;
 
-        uint startIndex = 0;
-        for (uint i = 0; i < length; i++) {
+        uint256 startIndex = 0;
+        for (uint256 i = 0; i < length; i++) {
             if (data[i] != 0) {
                 startIndex = i;
                 break;
@@ -189,10 +154,10 @@ library Bytes {
     }
 
     function removeEndingZero(bytes memory data) internal pure returns (bytes memory) {
-        uint length = data.length;
+        uint256 length = data.length;
 
-        uint endIndex = 0;
-        for (uint i = length - 1; i >= 0; i--) {
+        uint256 endIndex = 0;
+        for (uint256 i = length - 1; i >= 0; i--) {
             if (data[i] != 0) {
                 endIndex = i;
                 break;
@@ -203,10 +168,10 @@ library Bytes {
     }
 
     function reverse(bytes memory inbytes) internal pure returns (bytes memory) {
-        uint inlength = inbytes.length;
+        uint256 inlength = inbytes.length;
         bytes memory outbytes = new bytes(inlength);
 
-        for (uint i = 0; i <= inlength - 1; i++) {
+        for (uint256 i = 0; i <= inlength - 1; i++) {
             outbytes[i] = inbytes[inlength - i - 1];
         }
