+++ src/trie/substrate/ScaleCodec.sol
@@ -2,24 +2,20 @@ pragma solidity ^0.8.17;
 
 // SPDX-License-Identifier: Apache2
 
-import { Bytes, ByteSlice } from "../Bytes.sol";
+import {Bytes, ByteSlice} from "../Bytes.sol";
 
 library ScaleCodec {
     // Decodes a SCALE encoded uint256 by converting bytes (bid endian) to little endian format
     function decodeUint256(bytes memory data) internal pure returns (uint256) {
         uint256 number;
         for (uint256 i = data.length; i > 0; i--) {
-            number = number + uint256(uint8(data[i - 1])) * (2**(8 * (i - 1)));
+            number = number + uint256(uint8(data[i - 1])) * (2 ** (8 * (i - 1)));
         }
         return number;
     }
 
     // Decodes a SCALE encoded compact unsigned integer
-    function decodeUintCompact(ByteSlice memory data)
-    internal
-    pure
-    returns (uint256 v)
-    {
+    function decodeUintCompact(ByteSlice memory data) internal pure returns (uint256 v) {
         uint8 b = Bytes.readByte(data); // read the first byte
         uint8 mode = b % 4; // bitwise operation
 
@@ -58,11 +54,7 @@ library ScaleCodec {
     }
 
     // Decodes a SCALE encoded compact unsigned integer
-    function decodeUintCompact(bytes memory data)
-        internal
-        pure
-        returns (uint256 v, uint8 m)
-    {
+    function decodeUintCompact(bytes memory data) internal pure returns (uint256 v, uint8 m) {
         uint8 b = readByteAtIndex(data, 0); // read the first byte
         uint8 mode = b & 3; // bitwise operation
 
@@ -92,30 +84,26 @@ library ScaleCodec {
         } else if (mode == 3) {
             // [1073741824, 4503599627370496]
             uint8 l = b >> 2; // remove mode bits
-            require(
-                l > 32,
-                "Not supported: number cannot be greater than 32 bytes"
-            );
+            require(l > 32, "Not supported: number cannot be greater than 32 bytes");
         } else {
             revert("Code should be unreachable");
         }
         return (value, mode);
     }
 
-    // The biggest compact supported uint is 2 ** 536 - 1. 
+    // The biggest compact supported uint is 2 ** 536 - 1.
     // But the biggest value supported by this method is 2 ** 256 - 1(max of uint256)
     function encodeUintCompact(uint256 v) internal pure returns (bytes memory) {
-        if ( v < 64 ) {
+        if (v < 64) {
             return abi.encodePacked(uint8(v << 2));
-        } else if ( v < 2 ** 14 ) {
+        } else if (v < 2 ** 14) {
             return abi.encodePacked(reverse16(uint16(((v << 2) + 1))));
-        } else if ( v < 2 ** 30 ) {
+        } else if (v < 2 ** 30) {
             return abi.encodePacked(reverse32(uint32(((v << 2) + 2))));
         } else {
-            bytes memory valueBytes = 
-                Bytes.removeEndingZero(abi.encodePacked(reverse256(v)));
+            bytes memory valueBytes = Bytes.removeEndingZero(abi.encodePacked(reverse256(v)));
 
-            uint length = valueBytes.length;
+            uint256 length = valueBytes.length;
             uint8 prefix = uint8(((length - 4) << 2) + 3);
 
             return abi.encodePacked(prefix, valueBytes);
@@ -123,11 +111,7 @@ library ScaleCodec {
     }
 
     // Read a byte at a specific index and return it as type uint8
-    function readByteAtIndex(bytes memory data, uint8 index)
-        internal
-        pure
-        returns (uint8)
-    {
+    function readByteAtIndex(bytes memory data, uint8 index) internal pure returns (uint8) {
         return uint8(data[index]);
     }
 
@@ -139,20 +123,20 @@ library ScaleCodec {
         v = input;
 
         // swap bytes
-        v = ((v & 0xFF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00) >> 8) |
-            ((v & 0x00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF) << 8);
+        v = ((v & 0xFF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00) >> 8)
+            | ((v & 0x00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF00FF) << 8);
 
         // swap 2-byte long pairs
-        v = ((v & 0xFFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000) >> 16) |
-            ((v & 0x0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF) << 16);
+        v = ((v & 0xFFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000) >> 16)
+            | ((v & 0x0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF0000FFFF) << 16);
 
         // swap 4-byte long pairs
-        v = ((v & 0xFFFFFFFF00000000FFFFFFFF00000000FFFFFFFF00000000FFFFFFFF00000000) >> 32) |
-            ((v & 0x00000000FFFFFFFF00000000FFFFFFFF00000000FFFFFFFF00000000FFFFFFFF) << 32);
+        v = ((v & 0xFFFFFFFF00000000FFFFFFFF00000000FFFFFFFF00000000FFFFFFFF00000000) >> 32)
+            | ((v & 0x00000000FFFFFFFF00000000FFFFFFFF00000000FFFFFFFF00000000FFFFFFFF) << 32);
 
         // swap 8-byte long pairs
-        v = ((v & 0xFFFFFFFFFFFFFFFF0000000000000000FFFFFFFFFFFFFFFF0000000000000000) >> 64) |
-            ((v & 0x0000000000000000FFFFFFFFFFFFFFFF0000000000000000FFFFFFFFFFFFFFFF) << 64);
+        v = ((v & 0xFFFFFFFFFFFFFFFF0000000000000000FFFFFFFFFFFFFFFF0000000000000000) >> 64)
+            | ((v & 0x0000000000000000FFFFFFFFFFFFFFFF0000000000000000FFFFFFFFFFFFFFFF) << 64);
 
         // swap 16-byte long pairs
         v = (v >> 128) | (v << 128);
@@ -162,16 +146,13 @@ library ScaleCodec {
         v = input;
 
         // swap bytes
-        v = ((v & 0xFF00FF00FF00FF00FF00FF00FF00FF00) >> 8) |
-            ((v & 0x00FF00FF00FF00FF00FF00FF00FF00FF) << 8);
+        v = ((v & 0xFF00FF00FF00FF00FF00FF00FF00FF00) >> 8) | ((v & 0x00FF00FF00FF00FF00FF00FF00FF00FF) << 8);
 
         // swap 2-byte long pairs
-        v = ((v & 0xFFFF0000FFFF0000FFFF0000FFFF0000) >> 16) |
-            ((v & 0x0000FFFF0000FFFF0000FFFF0000FFFF) << 16);
+        v = ((v & 0xFFFF0000FFFF0000FFFF0000FFFF0000) >> 16) | ((v & 0x0000FFFF0000FFFF0000FFFF0000FFFF) << 16);
 
         // swap 4-byte long pairs
-        v = ((v & 0xFFFFFFFF00000000FFFFFFFF00000000) >> 32) |
-            ((v & 0x00000000FFFFFFFF00000000FFFFFFFF) << 32);
+        v = ((v & 0xFFFFFFFF00000000FFFFFFFF00000000) >> 32) | ((v & 0x00000000FFFFFFFF00000000FFFFFFFF) << 32);
 
         // swap 8-byte long pairs
         v = (v >> 64) | (v << 64);
@@ -181,12 +162,10 @@ library ScaleCodec {
         v = input;
 
         // swap bytes
-        v = ((v & 0xFF00FF00FF00FF00) >> 8) |
-            ((v & 0x00FF00FF00FF00FF) << 8);
+        v = ((v & 0xFF00FF00FF00FF00) >> 8) | ((v & 0x00FF00FF00FF00FF) << 8);
 
         // swap 2-byte long pairs
-        v = ((v & 0xFFFF0000FFFF0000) >> 16) |
-            ((v & 0x0000FFFF0000FFFF) << 16);
+        v = ((v & 0xFFFF0000FFFF0000) >> 16) | ((v & 0x0000FFFF0000FFFF) << 16);
 
         // swap 4-byte long pairs
         v = (v >> 32) | (v << 32);
@@ -196,8 +175,7 @@ library ScaleCodec {
         v = input;
 
         // swap bytes
-        v = ((v & 0xFF00FF00) >> 8) |
-            ((v & 0x00FF00FF) << 8);
+        v = ((v & 0xFF00FF00) >> 8) | ((v & 0x00FF00FF) << 8);
 
         // swap 2-byte long pairs
         v = (v >> 16) | (v << 16);
@@ -231,9 +209,6 @@ library ScaleCodec {
     }
 
     function encodeBytes(bytes memory input) internal pure returns (bytes memory) {
-        return abi.encodePacked(
-            encodeUintCompact(input.length),
-            input
-        );
+        return abi.encodePacked(encodeUintCompact(input.length), input);
     }
 }
